+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ADDED:
 
 - Allow middlewares to execute on "not found" routes.
+
+FIXED:
+
 - Make `--route-match` flag work for Middlewares. This flag was documented in
   the guide but wasn't registered in the command-line parsing logic.