+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased yet
+## 1.4.9
 
 ADDED: